templates: shortcodes: Replace about_company
Re-implement about_company(). New shortcode is named "kv_list" for Key
Value list. That's actually what it is.

The shortcode takes 4 args, three strings and one file name. The
strings are for left side strings.

The file name should point to a json file with key value pairs. The
shortcode prints all pairs in order. It can also take arbitrary HTML
snippets. So be careful when you add a new data.

CSS is not touched and still using old names.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>