templates: job_post: Add title block
Add a shared title block to align with other pages. The title block
should be factored out so that each template doesn't define its own,
but that's covered by a separate issue (#274).

Since an H1 was added, the open position name (title) is now an H2.

Although having an H3 above is not ideal, that's the current structure.
There's an open issue to fix this (#245).

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>